# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pasquali Solution - Gerenciamento de Colaboradores - Referência da API",
        "description": "Documentação dos endpoint disponíveis na API Restful",
        "contact": {
            "email": "randerccf@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Desenvolvimento(localhost)"
        },
        {
            "url": "https://api.testandoapi.click/api/v1",
            "description": "Produção"
        }
    ],
    "paths": {
...
        "/employees": {
            "get": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Listar Funcionários",
                "description": "Retorna a listagem de funcionários",
                "operationId": "4ce10d6079cc41f54345115536e6d623",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Employee"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthenticated. It's necessary to be authenticated to access this endpoint"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization Header": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Criar Funcionário",
                "description": "Cria um novo funcionário",
                "operationId": "11023790e5966928fb2d32057cab63a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "cpf",
                                    "rg",
                                    "birth",
                                    "user_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Alexandre da Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456-78"
                                    },
                                    "rg": {
                                        "type": "string",
                                        "example": "22.289.456-7"
                                    },
                                    "birth": {
                                        "type": "datetime",
                                        "example": "1981-09-16"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Employee"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "msg": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "O campo name é obrigatório!"
                                                },
                                                "cpf": {
                                                    "type": "string",
                                                    "example": "O campo cpf é obrigatório!"
                                                },
                                                "rg": {
                                                    "type": "string",
                                                    "example": "O campo rg é obrigatório!"
                                                },
                                                "birth": {
                                                    "type": "string",
                                                    "example": "O campo birth é obrigatório!"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "O campo email é obrigatório!"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "example": "O campo user_id é obrigatório!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthenticated. It's necessary to be authenticated to access this endpoint"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization Header": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "title": "Address",
                "description": "Classe Endereço",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "place": {
                        "type": "string",
                        "example": "Rua Maldonado"
                    },
                    "number": {
                        "type": "string",
                        "example": "694"
                    },
                    "city": {
                        "type": "string",
                        "example": "Dourados"
                    },
                    "state": {
                        "type": "string",
                        "example": "MT"
                    },
                    "postalCode": {
                        "type": "string",
                        "example": "67500-425"
                    },
                    "employee_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "created_at": {
                        "type": "datetime",
                        "example": "25/09/2021 23:43:47"
                    },
                    "updated_at": {
                        "type": "datetime",
                        "example": "25/09/2021 23:43:47"
                    }
                },
                "type": "object"
            },
            "Employee": {
                "title": "Employee",
                "description": "Classe Funcionário",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Alexandre da Silva"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "123.456-78"
                    },
                    "rg": {
                        "type": "string",
                        "example": "22.289.456-7"
                    },
                    "birth": {
                        "type": "datetime",
                        "example": "1981-09-16"
                    },
                    "email": {
                        "type": "string",
                        "example": "teste@example.com"
                    },
                    "currentSalary": {
                        "properties": {
                            "salary": {
                                "type": "number",
                                "example": "R$ 7.702,79"
                            }
                        },
                        "type": "object"
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Authorization Header": {
                "type": "http",
                "description": "Token de acesso JWT retornado após feito o login. Example: 'Authorization: Bearer {token}'",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}